@@ -5,8 +5,11 @@ class CharacterExpandableWidget extends StatelessWidget {
   final String header;
   final Widget child;
 
-  CharacterExpandableWidget({Key key, this.child, this.header})
-      : super(key: key);
+  CharacterExpandableWidget({
+    Key? key,
+    required this.child,
+    required this.header,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -15,6 +18,7 @@ class CharacterExpandableWidget extends StatelessWidget {
       child: Column(
         children: <Widget>[
           ExpandablePanel(
+            collapsed: Container(),
             header: Text(
               header,
               style: TextStyle(
