@@ -17,17 +17,19 @@ class SavedCharacterAdapter extends TypeAdapter<SavedCharacter> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return SavedCharacter(
-      characterId: fields[0] as int,
-      imageUrl: fields[1] as String,
-      name: fields[2] as String,
-      date: fields[4] as String,
+      characterId: fields[0] as int?,
+      imageUrl: fields[1] as String?,
+      name: fields[2] as String?,
+      date: fields[4] as String?,
+      anime: fields[5] as AnimeInformation?,
+      manga: fields[6] as AnimeInformation?,
     );
   }
 
   @override
   void write(BinaryWriter writer, SavedCharacter obj) {
     writer
-      ..writeByte(4)
+      ..writeByte(6)
       ..writeByte(0)
       ..write(obj.characterId)
       ..writeByte(1)
@@ -35,7 +37,11 @@ class SavedCharacterAdapter extends TypeAdapter<SavedCharacter> {
       ..writeByte(2)
       ..write(obj.name)
       ..writeByte(4)
-      ..write(obj.date);
+      ..write(obj.date)
+      ..writeByte(5)
+      ..write(obj.anime)
+      ..writeByte(6)
+      ..write(obj.manga);
   }
 
   @override
@@ -59,7 +65,7 @@ class KitsuDataAdapter extends TypeAdapter<KitsuData> {
     final fields = <int, dynamic>{
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
-    return KitsuData()..attributes = fields[0] as KitsuAttributes;
+    return KitsuData()..attributes = fields[0] as KitsuAttributes?;
   }
 
   @override
@@ -92,10 +98,10 @@ class KitsuAttributesAdapter extends TypeAdapter<KitsuAttributes> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return KitsuAttributes()
-      ..description = fields[0] as String
-      ..name = fields[1] as String
-      ..malId = fields[2] as int
-      ..image = fields[3] as KitsuImage;
+      ..description = fields[0] as String?
+      ..name = fields[1] as String?
+      ..malId = fields[2] as int?
+      ..image = fields[3] as KitsuImage?;
   }
 
   @override
@@ -133,7 +139,7 @@ class KitsuImageAdapter extends TypeAdapter<KitsuImage> {
     final fields = <int, dynamic>{
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
-    return KitsuImage()..original = fields[0] as String;
+    return KitsuImage()..original = fields[0] as String?;
   }
 
   @override
@@ -155,16 +161,51 @@ class KitsuImageAdapter extends TypeAdapter<KitsuImage> {
           typeId == other.typeId;
 }
 
+class AnimeInformationAdapter extends TypeAdapter<AnimeInformation> {
+  @override
+  final int typeId = 4;
+
+  @override
+  AnimeInformation read(BinaryReader reader) {
+    final numOfFields = reader.readByte();
+    final fields = <int, dynamic>{
+      for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
+    };
+    return AnimeInformation();
+  }
+
+  @override
+  void write(BinaryWriter writer, AnimeInformation obj) {
+    writer..writeByte(0);
+  }
+
+  @override
+  int get hashCode => typeId.hashCode;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is AnimeInformationAdapter &&
+          runtimeType == other.runtimeType &&
+          typeId == other.typeId;
+}
+
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
 SavedCharacter _$SavedCharacterFromJson(Map<String, dynamic> json) {
   return SavedCharacter(
-    characterId: json['characterId'] as int,
-    imageUrl: json['imageUrl'] as String,
-    name: json['name'] as String,
-    date: json['date'] as String,
+    characterId: json['characterId'] as int?,
+    imageUrl: json['imageUrl'] as String?,
+    name: json['name'] as String?,
+    date: json['date'] as String?,
+    anime: json['anime'] == null
+        ? null
+        : AnimeInformation.fromJson(json['anime'] as Map<String, dynamic>),
+    manga: json['manga'] == null
+        ? null
+        : AnimeInformation.fromJson(json['manga'] as Map<String, dynamic>),
   );
 }
 
@@ -174,13 +215,21 @@ Map<String, dynamic> _$SavedCharacterToJson(SavedCharacter instance) =>
       'imageUrl': instance.imageUrl,
       'name': instance.name,
       'date': instance.date,
+      'anime': instance.anime,
+      'manga': instance.manga,
     };
 
 JsonWaifu _$JsonWaifuFromJson(Map<String, dynamic> json) {
   return JsonWaifu(
-    title: json['title'] as String,
-    mal_id: json['mal_id'] as int,
-    image_url: json['image_url'] as String,
+    title: json['title'] as String?,
+    mal_id: json['mal_id'] as int?,
+    image_url: json['image_url'] as String?,
+    anime: json['anime'] == null
+        ? null
+        : AnimeInformation.fromJson(json['anime'] as Map<String, dynamic>),
+    manga: json['manga'] == null
+        ? null
+        : AnimeInformation.fromJson(json['manga'] as Map<String, dynamic>),
   );
 }
 
@@ -188,28 +237,24 @@ Map<String, dynamic> _$JsonWaifuToJson(JsonWaifu instance) => <String, dynamic>{
       'title': instance.title,
       'mal_id': instance.mal_id,
       'image_url': instance.image_url,
+      'anime': instance.anime,
+      'manga': instance.manga,
     };
 
 Waifu _$WaifuFromJson(Map<String, dynamic> json) {
   return Waifu(
-    name: json['name'] as String,
-    about: json['about'] as String,
-    image_url: json['image_url'] as String,
-    animeography: (json['animeography'] as List)
-        ?.map((e) => e == null
-            ? null
-            : AnimeInformation.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-    mangaography: (json['mangaography'] as List)
-        ?.map((e) => e == null
-            ? null
-            : AnimeInformation.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-    voice_actors: (json['voice_actors'] as List)
-        ?.map((e) => e == null
-            ? null
-            : AnimeInformation.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
+    name: json['name'] as String?,
+    about: json['about'] as String?,
+    image_url: json['image_url'] as String?,
+    animeography: (json['animeography'] as List<dynamic>?)
+        ?.map((e) => AnimeInformation.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    mangaography: (json['mangaography'] as List<dynamic>?)
+        ?.map((e) => AnimeInformation.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    voice_actors: (json['voice_actors'] as List<dynamic>?)
+        ?.map((e) => AnimeInformation.fromJson(e as Map<String, dynamic>))
+        .toList(),
   );
 }
 
@@ -224,15 +269,16 @@ Map<String, dynamic> _$WaifuToJson(Waifu instance) => <String, dynamic>{
 
 AnimeInformation _$AnimeInformationFromJson(Map<String, dynamic> json) {
   return AnimeInformation(
-    name: json['name'] as String,
-    image_url: json['image_url'] as String,
-    role: json['role'] as String,
-    language: json['language'] as String,
-  );
+    name: json['name'] as String?,
+    image_url: json['image_url'] as String?,
+    role: json['role'] as String?,
+    language: json['language'] as String?,
+  )..mal_id = json['mal_id'] as int?;
 }
 
 Map<String, dynamic> _$AnimeInformationToJson(AnimeInformation instance) =>
     <String, dynamic>{
+      'mal_id': instance.mal_id,
       'name': instance.name,
       'image_url': instance.image_url,
       'role': instance.role,
@@ -241,11 +287,9 @@ Map<String, dynamic> _$AnimeInformationToJson(AnimeInformation instance) =>
 
 FireBaseResponse _$FireBaseResponseFromJson(Map<String, dynamic> json) {
   return FireBaseResponse()
-    ..waifus = (json['waifus'] as List)
-        ?.map((e) => e == null
-            ? null
-            : SavedCharacter.fromJson(e as Map<String, dynamic>))
-        ?.toList();
+    ..waifus = (json['waifus'] as List<dynamic>?)
+        ?.map((e) => SavedCharacter.fromJson(e as Map<String, dynamic>))
+        .toList();
 }
 
 Map<String, dynamic> _$FireBaseResponseToJson(FireBaseResponse instance) =>
